requirements.txt
PyYAML==6.0
pyserial==3.5
pyautogui==0.9.53
opencv-python==4.7.0.72
pywin32==305
pygetwindow==0.0.9
numpy==1.24.2
